Fix printf segmentation fault
At 12:52 pm -0400 28/4/98, Ilya Zakharevich wrote:
>Hugo van der Sanden writes:
>> Using 5.004_05t1, the below gives an assertion failure:
>>   for ($i=0; $i<124; ++$i) {
>>       $data{$i}++;
>>   }
>>   foreach (0 .. 123) {
>>       @x = split(" ",undef);
>>       printf("%s%3d %s%6d\n",$x[0],$x[1],$x[2],$_);
>>   }
>> Run inside gdb with '-w /home/hsanden/t0', I get:
>>   Name "main::data" used only once: possible typo at /home/hsanden/t0 line 2.
>>   Use of uninitialized value at /home/hsanden/t0 line 5.
>>   assertion botched: *(unsigned int *)((caddr_t)Perl_op + Perl_op->ovu.ovu_size + 1 - sizeof (unsigned int)) == 0x55555555
>
>For those who have never seen it: This is a memory overrun.  Write was
>performed after an end of the malloced area.

Thanks Ilya - I hadn't seen it before.

I understand, I think, why it is happening: the $x[??] arguments to printf
are being compiled to pp_aelemfast, which pushes on the stack without
bounds checking - the second one exactly fills the stack, so the third
writes past the end; the final $_ is pp_gvsv, which does an EXTEND(sp, 1),
which grows the stack, copying all that it knows about, but leaving a
gap for the past-the-end value - this causes the SEGV when the overwrite
hasn't been trapped by the assertion.

I'm not sure whether AELEMFAST is supposed to EXTEND - if it is, the patch
below will suffice. If not, then presumably room on the stack is supposed
to be guaranteed at the time of generation of the op, which is in
op.c:peep() case OP_GV: I can't see there any attempt to check expected
stack depth.

I can certainly see the value of optimising the printf to something along
the lines of:
  EXTEND SP, 5
  MARK
  CONSTFAST
  AELEMFAST
  AELEMFAST
  AELEMFAST
  SCALARFAST
  PRINTF

.. but I suspect that isn't what 'FAST' is supposed to mean.

Patch is to 5.004_04, and passes all tests (and the above failure case)
here; it might be worth adding the failure case to op/misc.t if it is
at least reasonably likely to fail in the same place in the future - if
so, only the last four lines should be required.

p5p-msgid: l03130300b16bebdbc314@[194.222.64.89]