hpux.sh hints file clarification suggestion
Line 47 of the hpux.sh hints file indicates a work-around for
fixing a "too much defining problem". I would like to offer a
clarification to this for those who may not be aware that the
information listed is a SINGLE ccflag entry, not a list of entries
(separated by commas). The original and suggested clarifications
are list below:

ORIGINAL:
# If you get a message about "too much defining", you might have to
# add the following to your ccflags: '-Wp,-H256000'

# If you get a message about "too much defining", as may happen
# in HPUX < 10, you might have to append a single entry to your
# ccflags: '-Wp,-H256000'
# NOTE: This is a single entry (-W takes the argument 'p,-H256000').

Hopefully with the official end of HPUX 9.X support in Dec. 1998, 
this will no longer be much of a problem. You can't imagine how much
time I spent with variations of this ccflag before I got it to work.
This may be due to by newness to cc flags in 9.X, but who knows...

Thanks,
Matt Groener, mwg@qad.com

[Please do not change anything below this line]


p5p-msgid: 199802192351.QAA09096@jhereg.perl.com