Fixed Text/Wrap.pm bugs (2)
I recently found two bugs with Text::Wrap::wrap().

  A) Use of /m attribute
  ----------------------
  This causes long words to be moved to end of document.
  The example below shows a two test runs with old and new code.

  B) long word split and die
  --------------------------
  You currently still die if a long word on the first line exists.
  The long words split code only operates on subsequent lines.

After a bit of thought and a chat with Tim Bunce (on the other side of
the room) I decided to restructure the logic to simplify things a bit.
I believe that the output should not have changed for typical usage but
would appreciate more thorough testing.

[snip]

p5p-msgid: 199709291548.QAA08645@toad.ig.co.uk