another destruct_level fix
On Mon, 30 Mar 1998 23:47:55 +0200, Jan-Pieter Cornet wrote:
>Another fix to prevent coredumps when running with an increased
>destruct_level:
  [...]
>I believe that this fixes a typo or thinko of the original coder.

That's not accurate, I'm afraid.

>                                                                 Could the
>guilty party please stand up and confess in shame, or, in case I'm wrong,
>flame me for ruining their artistically crafted code and present the
>proper fix while doing so?

The HeVAL() slot of the string table contains a reference count (which
is not an IV for efficiency reasons).  The right value to signify a
reference count of zero is Nullsv, not &sv_undef.

At the time that code was written, hv_free_ent() didn't expect
HeVAL(entry) to contain a valid SV.  Someone changed that assumption.

>                          (I know that HeVAL of a HE in the strtab shouldn't
>contain an SV*, but hv_free_ent can't stand non-SV* values there, and
>I didn't know a better fix. This probably leaks hek_key[]s, but then again,
>that's the whole purpose of this code, to detect those leaks).

The "right" fix is to patch hv_free_ent() with:

p5p-msgid: 199804030105.UAA04400@aatma.engin.umich.edu