Fix constant detection in t/op/ipcsem.t for Digit UNIX
(sent also to perlbug and Tim, not sent to p5p at first because of a typo...)

As I said earlier the look-for-cpp-defines testing in t/op/ipc*.t was
broken, doomed, DOOMED.  For example the #include logic was broken,
twice: (1) something called $mm was the name of the #included file
(2) filehandle F was not local()ized.  Because of this much breakage
in Digital UNIX the S_IRWX[UGO] could not be found because they are
not directly in <sys/stat.h>.

IMNSHO The Right Way would be to create SysV::IPC (or IPC::SysV, whatever)
that contains the right constants, a la Fcntl.  I already took a shot at
this and found that one must sample several systems to garner all the IPC*,
SHM* SEM*, GET*, et alia constants.  If somebody feels like they know enough
of SVIPC (I don't) they can have what I've got scraped together.

However: I pushed the doom a little bit farther.  With this patch
Digital UNIX 4.0D is happy with 04-m2.  Mucho diagnostic output related
to the #include groveling added, they might come in handy when this
rickety house of cards falls down the next time.  When, not if.

p5p-msgid: 199805121212.PAA15351@alpha.hut.fi