MM_VMS.pm fixes for building external library
(I originally sent this out a while ago (back in january), but it didn't make it into the _05 patchkit. Here it is again)

Okay, folks. Here's a patch for MM_VMS.PM that fixes a bug when building
an extension with an external library. (like, say, SDBM_File or GD) It's not
perfect (will still fail under certain circumstances, but fewer than the
current version, and all currently functional cases still work) but it's good
enough to fix a majority of the cases *and* get SDBM_File to build (whith the
SDBM_File patches needed for VMS, of course)

p5p-msgid: 3.0.5.32.19980511160542.009dd480@ous.edu