Further VMS piping fixes from Charles Lane:
In summary, error messages produced when a subprocess terminated
abnormally were being sent not just to the parent process, but to
grandparents, because of default values for error output that were
not completely overridden when the subprocess was started.

This patch fixes this behavior by defining user-mode (i.e., temporary
for the duration of the program) logical names for SYS$OUTPUT and
SYS$ERROR when they are (re)opened inside Perl.  And a bunch of other
changes to make it so that the user-mode logicals are the ones that
control where Perl's error messages go if it terminates abnormally.

I also added some gratuitous fixes to the indentation of braces in
the piping code.  It just looked ugly, before.

p4raw-id: //depot/perl@8257