A modified version of
Subject: [ID 20010307.005] POSIX::sigaction has various problems
Message-Id: <200103072235.OAA25368@dt-sj1-130.sj.broadcom.com>

Currently the sigaction.t test #6 fails (and is fudged to look
like an "ok") in Linux (at least in Debian 2.2 Linux 2.4.2 x86).
This may well be a genuine bug in Linux sigaction() (since at
least Tru64, Solaris, and HP-UX disagree with Linux).
Anyone with POSIX / SUSv2 tome handy?  The problem is that
the flags of the oldaction don't match with the flags in
the previously installed disposition.

p4raw-id: //depot/perl@9086