The first steps towards cross-compilation.
Abstract execution of compiled test executables with $run,
and abstract transfer of files with $to and $from.

Under cross-compilation the $run, $to, and $from will point
to appropriate wrapper scripts, by default ssh and scp,
but also rsh, rcp, and cp are supported.  If not
cross-compiling, they will be set to '', ':', and ':',
respectively.

With these patches I was able to get Configure for
iPAQ ARM Linux on an Intel Linux about 95% right
(only a few tests failed to execute or they produced
incorrect results), and I was able to compile
a functional miniperl.

The symbol crosscompile renamed to be usecrosscompile,
the corresponding C symbol from CROSSCOMPILE to
USE_CROSS_COMPILE.

p4raw-id: //depot/perl@10592