Fix static extensions when building with nmake on Win32
They were broken because change 24806 forgot to add -DWITH_STATIC
to the recipe for compiling perllib.c which meant that perllibst.h
was not included and hence although they built OK, the static
extensions could not actually be used!

In fact, since perllibst.h is always created (it is more or less
empty when static extensions are not being used), there is no need
for -DWITH_STATIC at all. So rather than adding it to Makefile,
just drop it from makefile.mk and perllib.c.

Also add the missing cleanup of perllibst.h to Makefile.

p4raw-id: //depot/perl@25150