Fix $Config{ccflags} for Win32 perls built with dmake
The nmake makefile (win32/Makefile) currently adds -nologo, $(STRPOOL)
and -W3 to the ccflags in CFG_VARS, but the dmake makefile
(win32/makefile.mk) doesn't.  This means that extensions built with
perls built with VC++ and dmake spit out damn M$ logo messages and use
a lower warning level.

The reason for the difference is probably that the dmake makefile
supports multiple compilers.  Therefore, put the extra ccflags into a
new macro, and set it appropriately for each compiler.  Add the new
macro to the nmake makefile too for orthogonality.

p4raw-id: //depot/perl@25694