Sort the ext/ and lib/ tests when running under t/harness
Unless they are sorted then they are run in the same order in which
they are listed in MANIFEST, which is not always ideal. In particular,
the ext/Compress/IO/Zlib/t/*.t tests are not run in the correct order,
which causes some files to be left behind afterwards.

ExtUtils::Command::MM::test_harness() sorts test files, so it seems
sensible for t/harness to do likewise, rather than relying on the
ordering in MANIFEST.

p4raw-id: //depot/perl@27420