Correct large files / lseek size & type settings in Win32 canned configs.
Change 25208 switched off USE_LARGE_FILES in win32/config_H.* but left
LSEEKSIZE/Off_t_size and Off_t as 8 and __int64 (or long long)
respectively. Similarly change 25215 switched off uselargefiles in
win32/config.* but left lseeksize and lseektype as 8 and __int64 (or
long long) respectively. Change 25216 fixed the Borland settings in
win32/config.bc on the basis that Borland should always be using 4 and
long, but really all the other files should be using 4 and long for
their default values as well to match the default values of
USE_LARGE_FILES and uselargefiles. Having done that, we must then
reverse the logic for fiddling with these values in win32/config_sh.PL:
they are now changed to 8 and __int64 (or long long) if uselargefiles
*is* defined (except for Borland, which always wants 4 and long).

p4raw-id: //depot/perl@33347