Check for clz clzl availability
This defines PERL_USE_CLZ() to be the proper function to call on a
system to execute the equivalent of the count leading zero instruction
on the widest machine word.  It is left undefined if it couldn't find an
appropriate call.

clzll also exists, for long long, but I don't think we have long longs
different from plain long on platforms besides Windows, which doesn't
have this anyway.