Create single_1bit_pos()
Given a word known to have exactly a single bit set, this function uses
deBruijn sequences to quickly calculate its position.

I took the 32 and 64 bit word versions from the internet.  They differ
in how they treat a word with no set bits.  But this is considered
undefined behavior, so that difference is immaterial.

Apparently figuring this out uses brute force methods, and so I decided
to live with this difference, rather than to expend the time needed to
bring them into sync.