Create and use my_msbit_pos()
This inline function finds the bit position of the most significant 1
bit in a word.  It basically separates out code from another inline
function, so that it can be called more generally.

I took the liberty of clarifying the comment and explicitly #ifdef'ing
out a line of code instead of relying on the optimizer to skip it.