Use 'count leading zeros', if have it
This changes two functions to use clz if available.  On platforms with
that instruction, the expansion is likely to just it and a subtract,
beating the deBruijn method previously used.

And if that instruction isn't present, the libc emulation is likely to
be as fast as possible, again beating the hand-rolled deBruijn method.