Use ffs(), ffsl() for lsb_pos, if available
ffs() is a POSIX function that serves to find the lsb of a word.  It
should be faster than anything we hand-roll, but slower than any
available intrinsics that the previous commits use; so this commit looks
for it only if they aren't available.

Some platforms have an equivalent ffsl() for longs, which we use if the
word is long-sized.