Always use fast available msb/lsb method
Some platforms have a fast way to get the msb but not the lsb; others,
more rarely, have the reverse.  But using a few shift and the like
instructions allows us to reduce each instance to terms of the other.

This commit causes any available fast method to be used by turning the
non-available case into that one.